+++ web-wallet/src/lib/constants/documentationLinks.js
@@ -1,3 +1,4 @@
 export const DOCUMENTATION_LINKS = {
   RESTORE_BLOCK_HEIGHT: "https://docs.dusk.network/learn/guides/restore-height",
+  RUN_A_PROVISIONER: "https://docs.dusk.network/operator/02-provisioner/",
 };