+++ web-wallet/src/lib/stores/stores.d.ts
@@ -18,7 +18,6 @@ type SettingsStoreContent = {
   gasPrice: bigint;
   hideStakingNotice: boolean;
   language: string;
-  minAllowedStake: number;
   userId: string;
 };