+++ web-wallet/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Suggested words in the Mnemonic (Authenticate) are accessible using Tab [#2879]
 - Broken link in the stake warning [#2990]
+- Fix Dashboard navigation menu padding [#3000]
 
 ## [0.7.0] - 2024-11-11
 
@@ -341,6 +342,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2920]: https://github.com/dusk-network/rusk/issues/2920
 [#2932]: https://github.com/dusk-network/rusk/issues/2932
 [#2990]: https://github.com/dusk-network/rusk/issues/2990
+[#3000]: https://github.com/dusk-network/rusk/issues/3000
 
 <!-- VERSIONS -->