+++ web-wallet/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Suggested words in the Mnemonic (Authenticate) are accessible using Tab [#2879]
+- Enhance Allocate flow on full amount allocation [#2938]
 - Broken link in the stake warning [#2990]
 - Fix Dashboard navigation menu padding [#3000]
 
@@ -341,6 +342,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2888]: https://github.com/dusk-network/rusk/issues/2888
 [#2920]: https://github.com/dusk-network/rusk/issues/2920
 [#2932]: https://github.com/dusk-network/rusk/issues/2932
+[#2938]: https://github.com/dusk-network/rusk/issues/2938
 [#2990]: https://github.com/dusk-network/rusk/issues/2990
 [#3000]: https://github.com/dusk-network/rusk/issues/3000